implementando atualização de registro

+++ src/route.js
@@ -1,19 +1,23 @@
 import { Database } from "./database.js";
 import { randomUUID } from "node:crypto";
-import { buildRoutePath } from "./build-route-path.js";
+import { buildRoutePath } from "./utils/build-route-path.js";
 
 const database = new Database();
+const today = new Date().toLocaleString("pt-br");
 
 export const routes = [
   {
     method: "GET",
     path: buildRoutePath("/tasks"),
     handler: (req, res) => {
-      const tasks = database.select("tasks");
+      const { search } = req.query;
 
-      console.log(tasks);
+      const tasks = database.select("tasks", {
+        title: search,
+        description: search,
+      });
 
-      return res.writeHead(200).end(JSON.stringify(tasks));
+      return res.end(JSON.stringify(tasks));
     },
   },
   {
@@ -24,7 +28,17 @@ export const routes = [
 
       const { title, description } = req.body;
 
-      const today = new Date().toLocaleString("pt-br");
+      if (!title) {
+        return res
+          .writeHead(400)
+          .end(JSON.stringify({ message: "Titulo é obrigatório" }));
+      }
+
+      if (!description) {
+        return res
+          .writeHead(400)
+          .end(JSON.stringify({ message: "Descrição é obrigatório" }));
+      }
 
       const newTask = {
         id: randomUUID(),
@@ -44,8 +58,16 @@ export const routes = [
     path: buildRoutePath("/tasks/:id"),
     handler: (req, res) => {
       const { id } = req.params;
-      console.log(req.params);
-      database.delete('tasks', id)
+
+      const [task] = database.select("tasks", { id });
+
+      if (!task) {
+        return res
+          .writeHead(404)
+          .end(JSON.stringify({ message: "Tarefa não encontrada" }));
+      }
+
+      database.delete("tasks", id);
       return res.writeHead(204).end();
     },
   },
@@ -53,7 +75,25 @@ export const routes = [
     method: "PUT",
     path: buildRoutePath("/tasks/:id"),
     handler: (req, res) => {
-      return res.end();
+      const { id } = req.params;
+      const { title, description } = req.body;
+
+      if (!title || !description) {
+        return res
+          .writeHead(400)
+          .end(JSON.stringify({ message: "Título dou descrição obrigatório" }));
+      }
+
+      const [task] = database.select("tasks", { id });
+
+      if (!task) {
+        return res
+          .writeHead(404)
+          .end(JSON.stringify({ message: "Tarefa não encontrada" }));
+      }
+
+      database.update("tasks", id, { title, description, updated_at: today });
+      return res.writeHead(204).end();
     },
   },
   {
